+++ directives/extended-table.js
@@ -32,6 +32,7 @@
  *     data-action-list="actionList"
  *     data-cell-renderer="cellRenderer"
  *     data-column-class-callback="columnClassCallback"
+ *     data-table-class-callback="tableClassCallback"
  *     data-pager="pager"
  *     data-search="search"
  *     data-search-results="searchResults"&gt;&lt;/div&gt;
@@ -75,7 +76,7 @@
  *             '__keta.directives.ExtendedTable_no_entries': 'Pas d’entrées'
  *         };
  *
- *         // array of disabled components (empty by default)
+ *         // array of disabled components (default: everything except the table itself is disabled)
  *         $scope.disabledComponents = [
  *             // the table itself
  *             ExtendedTableConstants.COMPONENT.TABLE,
@@ -186,6 +187,11 @@
  *                 }
  *             }
  *             return columnClass;
+ *         };
+ *
+ *         // callback method to return class array for table
+ *         $scope.tableClassCallback = function() {
+ *             return ['table-striped'];
  *         };
  *
  *         // object for pager configuration (total, limit, offset)
@@ -320,6 +326,9 @@ angular.module('keta.directives.ExtendedTable',
 				// callback method to return class attribute for each column
 				columnClassCallback: '=?',
 
+				// callback method to return class array for table
+				tableClassCallback: '=?',
+
 				// object for pager configuration (total, limit, offset)
 				pager: '=?',
 
@@ -412,6 +421,11 @@ angular.module('keta.directives.ExtendedTable',
 				scope.columnClassCallback = scope.columnClassCallback || function() {
 					// parameters: row, column, isHeader
 					return '';
+				};
+
+				// tableClassCallback
+				scope.tableClassCallback = scope.tableClassCallback || function() {
+					return ['table-striped'];
 				};
 
 				// pager
@@ -573,6 +587,13 @@ angular.module('keta.directives.ExtendedTable',
 				}, true);
 
 				// ACTIONS ---
+
+				$scope.getTableClasses = function() {
+					var configuredClasses = $scope.tableClassCallback();
+					configuredClasses.push('table');
+					configuredClasses.push('form-group');
+					return configuredClasses.join(' ');
+				};
 
 				$scope.isDisabled = function(key) {
 					return inArray($scope.disabledComponents, key);
@@ -779,7 +800,7 @@ angular.module('keta.directives.ExtendedTable')
 '	<div class="row" data-ng-show="!isDisabled(COMPONENTS_TABLE)">' +
 '		<div class="col-xs-12">' +
 '			<div class="table-responsive table-data">' +
-'				<table class="table table-striped form-group">' +
+'				<table data-ng-class="getTableClasses()">' +
 '					<thead>' +
 '						<tr>' +
 '							<th class="{{columnClassCallback(headers, column, true)}} sortable"' +